+++ src/components/Icon/icons/IconPlay/index.js
@@ -1,0 +1,18 @@
+// @flow
+import * as React from 'react';
+import { css, cx } from 'emotion';
+import { Icon, type GenericIconProps } from '../../Icon';
+import image from './play.svg';
+
+const styles = css`
+  width: 16px;
+  height: 16px;
+`;
+
+export const IconPlay = ({ className, ...props }: GenericIconProps) => (
+  <Icon
+    {...props}
+    className={cx(styles, className)}
+    glyph={image}
+  />
+);